refactor(results): decompose ResultsDashboard into sub-components
Extracts three self-contained sections into src/components/results/:
- StepWarningsPanel: warning/error panel with per-step messages
- ComponentSummaryCards: clickable summary card grid
- ComponentBrowser: tabbed component browser (TabList + content)

ResultsDashboard is now a thin orchestrator managing only top-level
layout, main tab switching, and export dialog state.
