+++ src/components/ResultsDashboard.tsx
@@ -8,7 +8,6 @@ import {
   Card,
   Tab,
   TabList,
-  Tooltip,
   makeStyles,
   tokens,
   SelectTabData,
@@ -18,8 +17,6 @@ import {
   CheckmarkCircle24Regular,
   ArrowDownload24Regular,
   ArrowLeft24Regular,
-  Warning24Regular,
-  ErrorCircle24Regular,
 } from '@fluentui/react-icons';
 import {
   DashboardIcon,
@@ -39,7 +36,10 @@ import { SolutionDistributionView } from './SolutionDistributionView';
 import { ExportDialog } from './ExportDialog';
 import { FetchDiagnosticsView } from './FetchDiagnosticsView';
 import { Footer } from './Footer';
-import { COMPONENT_TABS, getDefaultTabKey } from './ComponentTabRegistry';
+import { getDefaultTabKey } from './ComponentTabRegistry';
+import { StepWarningsPanel } from './results/StepWarningsPanel';
+import { ComponentSummaryCards } from './results/ComponentSummaryCards';
+import { ComponentBrowser } from './results/ComponentBrowser';
 
 const useStyles = makeStyles({
   container: {
@@ -66,7 +66,6 @@ const useStyles = makeStyles({
     alignItems: 'center',
     marginBottom: tokens.spacingVerticalL,
   },
-  backButton: {},
   header: {
     marginBottom: tokens.spacingVerticalL,
     display: 'flex',
@@ -101,73 +100,6 @@ const useStyles = makeStyles({
   },
   section: {
     marginBottom: tokens.spacingVerticalL,
-  },
-  summaryGrid: {
-    display: 'grid',
-    gridTemplateColumns: 'repeat(auto-fit, minmax(140px, 1fr))',
-    gap: tokens.spacingHorizontalS,
-    marginTop: tokens.spacingVerticalS,
-  },
-  summaryCard: {
-    padding: tokens.spacingVerticalS,
-  },
-  summaryCardDisabled: {
-    padding: tokens.spacingVerticalS,
-    opacity: 0.5,
-    cursor: 'default',
-  },
-  summaryCardSelected: {
-    padding: tokens.spacingVerticalS,
-    borderBottom: `3px solid ${tokens.colorBrandForeground1}`,
-  },
-  summaryCardContent: {
-    display: 'flex',
-    flexDirection: 'column',
-    gap: tokens.spacingVerticalXXS,
-    alignItems: 'center',
-    textAlign: 'center',
-  },
-  summaryCount: {
-    fontSize: tokens.fontSizeHero800,
-    fontWeight: tokens.fontWeightSemibold,
-    lineHeight: tokens.lineHeightHero800,
-  },
-  summaryLabel: {
-    fontSize: tokens.fontSizeBase100,
-    color: tokens.colorNeutralForeground2,
-  },
-  warningsContainer: {
-    display: 'flex',
-    flexDirection: 'column',
-    gap: tokens.spacingVerticalM,
-  },
-  warningsPanel: {
-    display: 'flex',
-    flexDirection: 'column',
-    gap: tokens.spacingVerticalS,
-    padding: tokens.spacingVerticalM,
-    borderRadius: tokens.borderRadiusMedium,
-    border: `1px solid ${tokens.colorStatusWarningBorderActive}`,
-    backgroundColor: tokens.colorStatusWarningBackground1,
-  },
-  warningsPanelError: {
-    border: `1px solid ${tokens.colorStatusDangerBorderActive}`,
-    backgroundColor: tokens.colorStatusDangerBackground1,
-  },
-  warningRow: {
-    display: 'flex',
-    alignItems: 'flex-start',
-    gap: tokens.spacingHorizontalS,
-  },
-  warningStep: {
-    fontSize: tokens.fontSizeBase200,
-    color: tokens.colorNeutralForeground2,
-    minWidth: '110px',
-    fontWeight: tokens.fontWeightSemibold,
-  },
-  warningMessage: {
-    fontSize: tokens.fontSizeBase200,
-    color: tokens.colorNeutralForeground2,
   },
   browserSection: {
     marginTop: tokens.spacingVerticalXL,
@@ -209,16 +141,22 @@ export function ResultsDashboard({ result, scope, onStartOver }: ResultsDashboar
     }
   };
 
+  const handleCardClick = (key: string): void => {
+    setSelectedCard(key);
+    setSelectedTab(key);
+    browserSectionRef.current?.scrollIntoView({ behavior: 'smooth' });
+  };
+
+  const handleTabSelect = (key: string): void => {
+    setSelectedTab(key);
+    setSelectedCard(key);
+  };
+
   return (
     <div className={styles.container}>
       {/* Top Bar */}
       <div className={styles.topBar}>
-        <Button
-          className={styles.backButton}
-          appearance="subtle"
-          icon={<ArrowLeft24Regular />}
-          onClick={onStartOver}
-        >
+        <Button appearance="subtle" icon={<ArrowLeft24Regular />} onClick={onStartOver}>
           Generate New Blueprint
         </Button>
       </div>
@@ -247,34 +185,9 @@ export function ResultsDashboard({ result, scope, onStartOver }: ResultsDashboar
       </div>
 
       {/* Step warnings */}
-      {result.stepWarnings && result.stepWarnings.length > 0 && (() => {
-        const hasFullFailures = result.stepWarnings!.some(w => !w.partial);
-        return (
-          <div className={`${styles.warningsPanel}${hasFullFailures ? ` ${styles.warningsPanelError}` : ''}`}>
-            <div style={{ display: 'flex', alignItems: 'center', gap: tokens.spacingHorizontalS }}>
-              {hasFullFailures
-                ? <ErrorCircle24Regular style={{ color: tokens.colorStatusDangerForeground1, flexShrink: 0 }} />
-                : <Warning24Regular style={{ color: tokens.colorStatusWarningForeground1, flexShrink: 0 }} />
-              }
-              <Text weight="semibold" style={{ color: hasFullFailures ? tokens.colorStatusDangerForeground1 : tokens.colorStatusWarningForeground1 }}>
-                {hasFullFailures ? 'Some components could not be loaded' : 'Some data may be incomplete'}
-              </Text>
-              <Badge color="danger" shape="rounded" size="small" style={{ marginLeft: 'auto' }}>
-                {result.stepWarnings!.length} {result.stepWarnings!.length === 1 ? 'issue' : 'issues'}
-              </Badge>
-            </div>
-            {result.stepWarnings!.map((w, i) => (
-              <div key={i} className={styles.warningRow}>
-                <Text className={styles.warningStep}>{w.step}</Text>
-                <Text className={styles.warningMessage}>{w.message}</Text>
-              </div>
-            ))}
-            <Text style={{ fontSize: tokens.fontSizeBase200, color: tokens.colorNeutralForeground3 }}>
-              Open the <strong>Fetch Log</strong> tab for full API call details.
-            </Text>
-          </div>
-        );
-      })()}
+      {result.stepWarnings && result.stepWarnings.length > 0 && (
+        <StepWarningsPanel stepWarnings={result.stepWarnings} />
+      )}
 
       {/* Main Tabs */}
       <div className={styles.mainTabsSection}>
@@ -323,84 +236,23 @@ export function ResultsDashboard({ result, scope, onStartOver }: ResultsDashboar
           <div className={styles.section}>
             <Card>
               <Title3>Component Summary</Title3>
-              <div className={styles.summaryGrid}>
-                {COMPONENT_TABS.map((tab) => {
-                  const count = tab.count(result);
-                  const hasData = count > 0;
-                  const isSelected = selectedCard === tab.key;
-
-                  return (
-                    <Card
-                      key={tab.key}
-                      className={
-                        !hasData
-                          ? styles.summaryCardDisabled
-                          : isSelected
-                          ? styles.summaryCardSelected
-                          : styles.summaryCard
-                      }
-                      appearance={hasData ? 'filled' : 'outline'}
-                      onClick={
-                        hasData
-                          ? () => {
-                              setSelectedCard(tab.key);
-                              setSelectedTab(tab.key);
-                              browserSectionRef.current?.scrollIntoView({ behavior: 'smooth' });
-                            }
-                          : undefined
-                      }
-                      style={hasData ? { cursor: 'pointer' } : undefined}
-                    >
-                      <div className={styles.summaryCardContent}>
-                        <span style={{ color: hasData ? tokens.colorBrandForeground1 : tokens.colorNeutralForeground4, lineHeight: 1 }}>{tab.icon}</span>
-                        <Text className={styles.summaryCount}>{count}</Text>
-                        <Text className={styles.summaryLabel}>{tab.label}</Text>
-                      </div>
-                    </Card>
-                  );
-                })}
-              </div>
+              <ComponentSummaryCards
+                result={result}
+                selectedCard={selectedCard}
+                onCardClick={handleCardClick}
+              />
             </Card>
           </div>
 
-          {/* SECTION 3: Component Browser (Tabbed Interface) */}
+          {/* SECTION 3: Component Browser */}
           <div className={styles.browserSection} ref={browserSectionRef}>
             <Card>
               <Title3>Component Browser</Title3>
-              <TabList
-                selectedValue={selectedTab}
-                onTabSelect={(_event: SelectTabEvent, data: SelectTabData) => {
-                  setSelectedTab(data.value as string);
-                  setSelectedCard(data.value as string);
-                }}
-                size="small"
-                style={{
-                  flexWrap: 'wrap',
-                  gap: tokens.spacingHorizontalS,
-                }}
-              >
-                {COMPONENT_TABS.map((tab) => {
-                  if (tab.hidden?.(result)) return null;
-                  const count = tab.count(result);
-                  const isSelected = selectedTab === tab.key;
-                  return (
-                    <Tooltip key={tab.key} content={tab.label} relationship="label">
-                      <Tab value={tab.key} icon={tab.icon}>
-                        {isSelected ? `${tab.label} (${count})` : `${count}`}
-                      </Tab>
-                    </Tooltip>
-                  );
-                })}
-              </TabList>
-
-              {/* Tab Content — driven by registry */}
-              <div style={{ marginTop: tokens.spacingVerticalL }}>
-                {COMPONENT_TABS.map((tab) => {
-                  if (selectedTab !== tab.key) return null;
-                  if (tab.hidden?.(result)) return null;
-                  return <div key={tab.key}>{tab.render(result)}</div>;
-                })}
-              </div>
+              <ComponentBrowser
+                result={result}
+                selectedTab={selectedTab}
+                onTabSelect={handleTabSelect}
+              />
             </Card>
           </div>
         </>
